+++ src/algorithms/cocktail_shaker.ts
@@ -1,4 +1,4 @@
-import { SortHistory } from 'types';
+import { SortHistory } from 'shared/types';
 import { generateFromToArray, swapUnsafe } from 'lib/array';
 import {
   addToComparing,
