+++ src/components/visualizer/components/legend/Legend.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { LegendItem as Item } from 'types';
+import { LegendItem as Item } from 'shared/types';
 
 type LegendItemProps = Item & {
   id: string;
